+++ tests/FakeXrmEasy.Plugins.Tests/Pipeline/DefaultValuesInPipelineTests.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Linq;
 using DataverseEntities;
 using FakeXrmEasy.Abstractions.Plugins.Enums;
@@ -114,6 +115,44 @@ namespace FakeXrmEasy.Plugins.Tests.Pipeline
             var target = executedPluginSteps.First().TargetEntity as dv_test;
             Assert.False(target.dv_bool);
             Assert.Equal(id, target.Id);
+            Assert.Equal(_context.CallerProperties.CallerId.Id, target.CreatedBy.Id);
+            Assert.Equal(_context.CallerProperties.CallerId.LogicalName, target.CreatedBy.LogicalName);
+            Assert.Equal(_context.CallerProperties.CallerId.Id, target.ModifiedBy.Id);
+            Assert.Equal(_context.CallerProperties.CallerId.LogicalName, target.ModifiedBy.LogicalName);
+            Assert.Equal(_context.CallerProperties.CallerId.Id, target.OwnerId.Id);
+            Assert.Equal(_context.CallerProperties.CallerId.LogicalName, target.OwnerId.LogicalName);
+            Assert.Equal(_context.CallerProperties.CallerId.Id, target.OwningUser.Id);
+            Assert.Equal(_context.CallerProperties.CallerId.LogicalName, target.OwningUser.LogicalName);
+        }
+        
+        [Fact]
+        public void Should_not_override_attributes_in_target_entity_with_defaults_in_preoperation_step()
+        {
+            _context.InitializeMetadata(typeof(dv_test).Assembly);
+            
+            _context.RegisterPluginStep<TracerPlugin>(new PluginStepDefinition()
+            {
+                EntityLogicalName = dv_test.EntityLogicalName,
+                MessageName = "Create",
+                Stage = ProcessingStepStage.Preoperation,
+                Mode = ProcessingStepMode.Synchronous,
+            });
+
+            var guid = Guid.NewGuid();
+            var id = _service.Create(new dv_test()
+            {
+                Id = guid,
+                dv_bool = true
+            });
+            
+            var executedPluginSteps = _context.GetPluginStepAudit().CreateQuery().ToList();
+            Assert.Single(executedPluginSteps);
+
+            var target = executedPluginSteps.First().TargetEntity as dv_test;
+            
+            Assert.True(target.dv_bool);
+            Assert.Equal(id, target.Id);
+            Assert.Equal(guid, target.Id);
             Assert.Equal(_context.CallerProperties.CallerId.Id, target.CreatedBy.Id);
             Assert.Equal(_context.CallerProperties.CallerId.LogicalName, target.CreatedBy.LogicalName);
             Assert.Equal(_context.CallerProperties.CallerId.Id, target.ModifiedBy.Id);